refactor votes logic with new schema and logic
Find in here a module that is not being utilized. Cannot get to work, need help. Aside from that we are passing tutorial id thorugh the vote_path route now. Vote attribute for user_id has been changed to voter_id for understandability. Score methods and logic have been implemented in the ApplicationController. Need to get them out and into a module. Refactor is in order